Add Storage abstract class to decouple Logic class from StorageFile class
